Move WithBlockStateFailure to plt-scheduler
