+++ plt/plt-scheduler/src/scheduler/p11.rs
@@ -1,6 +1,6 @@
 use crate::scheduler::{ChainUpdateExecutionError, TransactionExecutionError, plt_scheduler};
 use crate::transaction_execution::{OutOfEnergyError, TransactionExecution};
-use crate::{TransactionContext, protocol_level_tokens};
+use crate::{TransactionContext, failure, protocol_level_tokens};
 use concordium_base::protocol_level_tokens::meta_operations::{
     LockOperation, MetaUpdateOperation, MetaUpdateOperations, MetaUpdatePayload,
 };
@@ -12,7 +12,7 @@ use plt_block_state::entity::accounts::Account;
 use plt_block_state::entity::block_state::p11::BlockStateP11;
 use plt_block_state::entity::{EntityContext, EntityContextTypes};
 use plt_block_state::failure::BlockStateResult;
-use plt_block_state::{failure, utils};
+use plt_block_state::utils;
 use plt_scheduler_types::types::execution::{
     ChainUpdateOutcome, TransactionExecutionSummary, TransactionOutcome,
 };
