+++ plt/plt-scheduler/tests/utils/entity_traits/scheduler_impl.rs
@@ -11,12 +11,11 @@ use plt_block_state::entity::block_state::TokenNotFoundByIdError;
 use plt_block_state::entity::block_state::p9::BlockStateP9;
 use plt_block_state::entity::block_state::p11::BlockStateP11;
 use plt_block_state::entity::{EntityContext, EntityContextTypes};
-use plt_block_state::failure;
 use plt_block_state::failure::BlockStateResult;
-use plt_scheduler::protocol_level_tokens;
 use plt_scheduler::queries::QueryLockError;
 use plt_scheduler::scheduler::{ChainUpdateExecutionError, TransactionExecutionError};
 use plt_scheduler::{TransactionContext, queries, scheduler};
+use plt_scheduler::{failure, protocol_level_tokens};
 use plt_scheduler_types::types::execution::{ChainUpdateOutcome, TransactionExecutionSummary};
 use plt_scheduler_types::types::queries::{TokenAccountInfo, TokenAuthorizations, TokenInfo};
 use std::mem;
